Bugfix toggleSelectAll() with groupBy() in query builder
If we use groupBy() in the queryBuilder, the count() method in toggleSelectAll() return a wrong value. As a result, the toggleSelectAll() doesn't work as expected selecting all checkboxes works, but deselecting all of them does not).

This is an already known issue : https://github.com/laravel/framework/issues/22883 and I use the approach https://github.com/laravel/framework/issues/22883#issuecomment-438521002 to fix the bug